separate stochastic policy net
